Adapt unit test to show bug
This small change to the H5LIterate test shows that when calling
H5L.iterate with a H5.index_t value other than H5.index_t.Name (which is
represented as 0) the parameter is not correctly translated by
HDF.PInvoke. For instance using H5.index_t.CRT_ORDER results in the
value 483766032 being passed to the native code, this value is than
rejected by the parameter checking in the native library.